requirements.txt
streamlit
requests
fastapi
uvicorn
python-dotenv
pydantic
langchain-core
langchain-community
langchain-google-genai
langchain-huggingface
langgraph
faiss-cpu
pypdf
huggingface-hub
transformers
torch
sentence-transformers
duckduckgo-search
doc2txt
unstructured[all-docs]
python-magic
pandas
openpyxl
xlrd